Bunch of updates and fixes. Pushing to repro so I can test things out.
