+++ plugin.xml
@@ -3,11 +3,12 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     id="org.thaliproject.p2p"
     version="0.0.1">
-    <name>ThaliPlugin</name>
+    <name>Thali</name>
     <description>ThaliPlugin to be used on top of JXcore</description>
     <license>MIT</license>
     <keywords>cordova,jxcore,node.js,thali</keywords>
-    <dependency id="io.jxcore.node" url="https://github.com/jxcore/jxcore-cordova" />
+    <dependency id="io.jxcore.node" url="https://github.com/jxcore/jxcore-cordova" 
+        commit="47e4eeb4b5a3131f47770259d9c02949069ec557" />
     <engines>
         <engine name="cordova-android" version=">=3.2.0" />
         <engine name="android-sdk" version=">=19" />
@@ -20,7 +21,6 @@
             <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
             <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
             <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
-            <uses-sdk android:minSdkVersion="19" android:targetSdkVersion="22" />
         </config-file>
         <framework src="src/android/JXcore.gradle" custom="true" type="gradleReference" />
         <source-file src="src/android/java/io/jxcore/node/BtToRequestSocket.java" target-dir="src/io/jxcore/node/" />
@@ -30,6 +30,7 @@
         <source-file src="src/android/java/io/jxcore/node/ConnectivityMonitor.java" target-dir="src/io/jxcore/node/" />
         <source-file src="src/android/java/io/jxcore/node/LifeCycleMonitor.java" target-dir="src/io/jxcore/node/" />
         <source-file src="src/android/java/io/jxcore/node/BtSocketDisconnectedCallBack.java" target-dir="src/io/jxcore/node/" />
+        <hook type="before_build" src="scripts/androidBeforeBuild.js"/>
     </platform>
 
     <!-- iOS -->
@@ -58,6 +59,4 @@
         <header-file src="src/ios/THEPeerNetworking.h" />
         <source-file src="src/ios/THEPeerNetworking.m" />
     </platform>
-    	<!-- Cordova hook for Thali pre-requisite -->
-		<hook type="after_plugin_add" src="scripts/thaliPostInstall.js" />
 </plugin>
